+++ cypress/integration/optionsPanel.spec.js
@@ -1,6 +1,5 @@
 /// <reference types="cypress" />
 
-// TODO - Stub these responses
 context("Options Panel", () => {
   beforeEach(() => {
     cy.fixture("jobs50").then((jobsJson) => {
@@ -16,18 +15,28 @@ context("Options Panel", () => {
           method: "GET",
           url: "/jobs/search?full_time=true&description=developer",
           status: 200,
+          delay: 1000,
           response: jobsSearch1Json,
         });
         cy.route({
           method: "GET",
           url: "/jobs/search?full_time=false&description=&location=Los Angeles",
           status: 200,
+          delay: 1000,
           response: jobsSearch1Json,
         });
         cy.route({
           method: "GET",
           url: "/jobs/search?full_time=false&description=&location=Chicago",
           status: 200,
+          delay: 1000,
+          response: jobsSearch1Json,
+        });
+        cy.route({
+          method: "GET",
+          url: "/jobs/search?full_time=false&description=developer",
+          status: 200,
+          delay: 1000,
           response: jobsSearch1Json,
         });
       });
@@ -68,5 +77,25 @@ context("Options Panel", () => {
     cy.get(".search__button").click();
     cy.wait(1000);
     cy.get("#location-1").should("be.checked");
+  });
+
+  it("Should filter with fullTime correctly", () => {
+    cy.get("#search").type("developer");
+    cy.get(".search__button").click();
+
+    cy.wait(1500);
+    cy.get(".jobcard__container").then(($jobs) => {
+      assert.equal($jobs.length, 5);
+    });
+
+    cy.get(
+      "#app > div.search__container > div.options-panel__container > label:nth-child(1) > span"
+    ).click();
+    cy.get(".search__button").click();
+
+    cy.wait(1500);
+    cy.get(".jobcard__container").then(($jobs) => {
+      assert.equal($jobs.length, 2);
+    });
   });
 });